Move aggregate `Registry` to the package level.
We want this class to be more visible as it plays important role in lifecycle of aggregates.
